Dynamic attributes are unset while saving a model

+++ app/Http/Controllers/Admin/ProductsController.php
@@ -124,29 +124,36 @@ class ProductsController extends Controller
         $updated = $product->update($request->validated());
 
         // UPDATES PRODUCT'S ATTRIBUTES
-        foreach (EntityType::where('label', Product::class)->first()->attributes()->where('is_system', false)->get() as $attribute) {
+        if ($request->input('attributes'))
+        {
+            Log::debug($product);
+            foreach (EntityType::where('label', Product::class)->first()->attributes()->where('is_system', false)->get() as $attribute) {
 
-            $value = $request->input('attributes')[$attribute->id] ?? null;
-
-            if (null === $value)
-            {
-                $product->attributes()->detach($attribute->id);
-
-                // TO DO EAV* value field is not deleted
-            } else {
-                $value_id = $attribute->type::findOrCreate($product, $attribute, $value);
-
-                $product->attributes()->syncWithoutDetaching([
-                    $attribute->id => [
-                        'value_type' => $attribute->type,
-                        'value_id' => $value_id,
-                    ],
-                ]);
+                $value = $request->input('attributes')[$attribute->id] ?? null;
+    
+                if (null === $value)
+                {
+                    $product->attributes()->detach($attribute->id);
+    
+                    // TO DO EAV* value field is not deleted
+                } else {
+                    $value_id = $attribute->type::findOrCreate($product, $attribute, $value);
+    
+                    $product->attributes()->syncWithoutDetaching([
+                        $attribute->id => [
+                            'value_type' => $attribute->type,
+                            'value_id' => $value_id,
+                        ],
+                    ]);
+                }
             }
         }
 
         // UPDATES CATEGORIES
-        $product->categories()->sync($request->input('categories'));
+        if ($request->input('categories'))
+        {
+            $product->categories()->sync($request->input('categories'));
+        }
 
         return response()->json([
             'updated' => $updated,
